Remove dependencies across modules
