+++ tests/tacacs/test_accounting.py
@@ -1,17 +1,11 @@
 import logging
 import time
-from tests.common.devices.ptf import PTFHost
-
-
 import pytest
-
-
-from .test_authorization import ssh_connect_remote_retry, ssh_run_command, \
-        remove_all_tacacs_server
+from tests.common.devices.ptf import PTFHost
 from tests.common.helpers.tacacs.tacacs_helper import stop_tacacs_server, start_tacacs_server, \
-    per_command_accounting_skip_versions
+    per_command_accounting_skip_versions, remove_all_tacacs_server
 from .utils import check_server_received, change_and_wait_aaa_config_update, get_auditd_config_reload_timestamp, \
-    ensure_tacacs_server_running_after_ut    # noqa: F401
+    ensure_tacacs_server_running_after_ut, ssh_connect_remote_retry, ssh_run_command    # noqa: F401
 from tests.common.errors import RunAnsibleModuleFail
 from tests.common.helpers.assertions import pytest_assert
 from tests.common.utilities import skip_release
@@ -22,7 +16,6 @@ pytestmark = [
     pytest.mark.topology('any'),
     pytest.mark.device_type('vs')
 ]
-
 
 logger = logging.getLogger(__name__)
